Second Project in - Analytics Programming Course
This project we will make use of publicly available data regarding Gun Violence Incidents in the United States, focusing on data related to Mass Shooting Incidents between 2014 and 2019.

The research questions for this project are:
What were the top 5 US States and Citys by Percentage of Total Mass Shooting incidents in the US over the period 2014-2019 ?
What was the distribution of total people involed (killed and injured) over that period in the city with the most incidents ?

The answers to these questions would be beneficial for both the private citizen and the government. Individuals who are considering relocating to a new state or city would want to be able to compare the frequency of gun violence occurrences between their current and possible future home.
Furthermore, all levels of government (local, state, and federal) should be interested in these answers. The outcomes should affect the amount of preventative action the government employs (police presence on the streets, gun control efforts, investments in community programs, and more).