Create Final_Project_Analyzing_CO2_Emissions_GDP_and_Energy_Consumption.ipynb
In this project, we will try to determine the extent to which each BRICS nation has contributed to Global Warming as opposed to the contribution of OECD Nations. "Contribution" will be in terms of consumption of traditionally generated energy as opposed to the consumption of renewable energy, and the amount of Carbon Dioxide (CO2) Emitted in BRICS nations which are nations with emerging economies and have a population of approximately 3 billion people7 – 40% of the world’s total population.

The research questions for this project will be:

1. In the past five years, what is the percentage of renewable energy consumption of the total energy consumption in the BRICS nations in comparison to the OECD nations?

The results of this research would emphasize which nations are "falling behind" the developed world standard. It should pressure them to implement new regulations that would promote energy efficiency, generation of renewable energy, and new technologies that would contribute to energy security and environmental and social benefits.

2. Looking at the yearly trend (from 1990 to 2018), what is the relationship between the GDP (per capita) and total energy consumption (per capita) in the BRICS nations?

3. What was the yearly CO2 emission level (per capita) in each of the BRICS nations between 1990 and 2018 compared to the OECD average? And what was the average CO2 emission level in the BRICS nations compared to the OECD average?